add quotes when parsing JSON parameters
Currently the program fails to process OCR's pages which contain the word "locale" in the OCR's text, because it interprets OCR'd text as locale definition! 
Searching for JSON parameter names in quotation signs with a colon should be safe, because in unescaped quotation signs followed by a colon are only used in JSON as parameter names.